+++ site/app/(marketing)/_components/footer.tsx
@@ -1,0 +1,103 @@
+import Image from "next/image";
+import { Mail, Phone, MapIcon } from "lucide-react";
+const navLinks = [
+  { href: "#about", label: "About" },
+  { href: "#services", label: "Services" },
+  { href: "#properties", label: "Properties" },
+];
+
+export function Footer() {
+  return (
+    <footer className="bg-gradient-to-br from-background via-accent/5 to-primary/5 border-t border-border/20">
+      <div className="max-w-7xl mx-auto px-4 py-16">
+        <div className="grid grid-cols-1 md:grid-cols-4 gap-8 mb-12">
+          {/* Company Info */}
+          <div className="md:col-span-2">
+            <div className="flex items-center gap-3 mb-4">
+              <Image
+                className="w-7 h-7 text-primary"
+                width={100}
+                height={100}
+                src="/logo.png"
+                alt="logo"
+              />
+              <div>
+                <span className="text-xl font-bold text-foreground">Atria</span>
+                <p className="text-xs  text-muted-foreground">
+                  Fractional Property Investment in Kenya
+                </p>
+              </div>
+            </div>
+            <p className="text-muted-foreground text-sm leading-relaxed mb-6 max-w-md">
+              Making real estate investment accessible to every Kenyan through
+              fractional ownership.
+            </p>
+          </div>
+
+          {/* Quick Links */}
+          <div>
+            <h3 className="font-semibold text-foreground mb-4">Quick Links</h3>
+            <ul className="space-y-2">
+              {navLinks.map((link) => (
+                <li key={link.href}>
+                  <a
+                    href={link.href}
+                    className="text-muted-foreground hover:text-primary transition-colors duration-200"
+                  >
+                    {link.label}
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </div>
+
+          {/* Contact Info */}
+          <div>
+            <h3 className="font-semibold text-foreground mb-4">Get in Touch</h3>
+            <div className="space-y-3">
+              <div className="flex items-center gap-3 text-muted-foreground">
+                <Mail className="w-4 h-4 text-primary" />
+                <span className="text-sm">roman.njoroge@njuguna.com</span>
+              </div>
+              <div className="flex items-center gap-3 text-muted-foreground">
+                <Phone className="w-4 h-4 text-primary" />
+                <span className="text-sm">+254 702 735922</span>
+              </div>
+              <div className="flex items-center gap-3 text-muted-foreground">
+                <MapIcon className="w-4 h-4 text-primary" />
+                <span className="text-sm">Nairobi, Kenya</span>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        {/* Bottom Bar */}
+        <div className="border-t border-border/20 pt-8 flex flex-col md:flex-row justify-between items-center gap-4">
+          <p className="text-sm text-muted-foreground">
+            © 2025 Atria. All rights reserved.
+          </p>
+          <div className="flex gap-6 text-sm text-muted-foreground">
+            <a
+              href="#"
+              className="hover:text-primary transition-colors duration-200"
+            >
+              Privacy Policy
+            </a>
+            <a
+              href="#"
+              className="hover:text-primary transition-colors duration-200"
+            >
+              Terms of Service
+            </a>
+            <a
+              href="#"
+              className="hover:text-primary transition-colors duration-200"
+            >
+              Support
+            </a>
+          </div>
+        </div>
+      </div>
+    </footer>
+  );
+}
